src/main/java/org/pantheon/gaia/compileur/outils/analyseur/lexique/vocabulaire/v1/ParentheseFermanteLexique.java
package org.pantheon.gaia.compileur.outils.analyseur.lexique.vocabulaire.v1;

import org.pantheon.gaia.compileur.outils.analyseur.lexique.vocabulaire.source.etat.EtatAutomateSatisfaisant;
import org.pantheon.gaia.compileur.outils.analyseur.lexique.vocabulaire.source.Lexique;
import org.pantheon.gaia.compileur.outils.analyseur.lexique.vocabulaire.source.etat.EtatAutomateNonSatisfaisant;
import org.pantheon.gaia.compileur.outils.analyseur.lexique.vocabulaire.source.transition.TransitionAutomate;
import org.pantheon.gaia.compileur.outils.symbole.source.Symbole;
import org.pantheon.gaia.compileur.outils.symbole.v1.ParentheseFermanteSymbole;

/**
 * Champ lexical pour les types de parenthèses fermantes.
 */
public class ParentheseFermanteLexique extends Lexique {

    /**
     * Initialise une nouvelle instance de la classe {@link ParentheseFermanteLexique}.
     */
    public ParentheseFermanteLexique() {
        super();
    }

    @Override
    protected void initialiser() {
        EtatAutomateNonSatisfaisant etat1 = new EtatAutomateNonSatisfaisant();
        EtatAutomateSatisfaisant etat2 = new EtatAutomateSatisfaisant();

        TransitionAutomate transition = new TransitionAutomate(etat2, "\\)");

        etat1.ajtTransition(transition);
        this.source = etat1;
    }

    @Override
    public Symbole extraireSymbole(String instruction, int curseur) {
        String unite = this.extraireUniteLexicale(instruction, curseur);
        if(unite == null) return null;
        return new ParentheseFermanteSymbole(unite);
    }
}
